Refactor diagonal into separate file
